// src/lib/enums/product.enum.ts
export enum ProductSize {
  SMALL = "SMALL",
  NORMAL = "NORMAL",
  LARGE = "LARGE",
  SET = "SET",
}

export enum ProductStatus {
  PAUSE = "PAUSE",
  PROCESS = "PROCESS",
  DELETE = "DELETE",
}

export enum ProductCollection {
  CLIMBING = "CLIMBING",
  CAMPING = "CAMPING",
  HIKING = "HIKING",
  TREKKING = "TREKKING",
  CYCLING = "CYCLING",
  APPAREL = "APPAREL",
  FOOTWEAR = "FOOTWEAR",
  OTHER = "OTHER",
}